Solid noise & basic terrain generation

+++ littleVulkanEngine/core/lve_app.cpp
@@ -72,7 +72,7 @@ void LveApp::createCommandBuffers() {
     throw std::runtime_error("failed to allocate command buffers!");
   }
 
-  for (size_t i = 0; i < commandBuffers_.size(); i++) {
+  for (int i = 0; i < commandBuffers_.size(); i++) {
     VkCommandBufferBeginInfo beginInfo = {};
     beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
     beginInfo.flags = 0;                   // Optional
